requirements.txt
PyQt6==6.10.2
qasync==0.27.1
httpx==0.28.1
pydantic==2.6.4
comicbox==2.2.1
Pillow==10.2.0
